Cleaner to do this from inside `Engine`
