+++ src/main/java/org/jenkinsci/remoting/engine/WorkDirManager.java
@@ -36,17 +36,12 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.PrintStream;
-import java.nio.channels.Channels;
-import java.nio.channels.FileChannel;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.StandardOpenOption;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.logging.FileHandler;
 import java.util.logging.Level;
 import java.util.logging.LogManager;
@@ -85,9 +80,6 @@ public class WorkDirManager {
      * Name of the standard System Property, which points to the {@code java.util.logging} config file.
      */
     public static final String JUL_CONFIG_FILE_SYSTEM_PROPERTY_NAME = "java.util.logging.config.file";
-
-    /** {@link #initializeWorkDir} is called twice. */
-    private Path internalDirPath;
 
     // Status data
     boolean loggingInitialized;
@@ -185,7 +177,6 @@ public class WorkDirManager {
      * @param failIfMissing Fail the initialization if the workDir or internalDir are missing.
      *                      This option presumes that the workspace structure gets initialized previously in order to ensure that we do not start up with a borked instance
      *                      (e.g. if a mount gets disconnected).
-     * @param agentName name of the agent
      * @return Initialized directory for internal files within workDir or {@code null} if it is disabled
      * @throws IOException Workspace allocation issue (e.g. the specified directory is not writable).
      *                     In such case Remoting should not start up at all.
@@ -193,15 +184,8 @@ public class WorkDirManager {
     @CheckForNull
     @SuppressFBWarnings(value = "PATH_TRAVERSAL_IN", justification = "Parameter supplied by user / administrator.")
     public Path initializeWorkDir(
-            final @CheckForNull File workDir,
-            final @NonNull String internalDir,
-            final boolean failIfMissing,
-            @NonNull String agentName)
+            final @CheckForNull File workDir, final @NonNull String internalDir, final boolean failIfMissing)
             throws IOException {
-
-        if (internalDirPath != null) {
-            return internalDirPath;
-        }
 
         if (!internalDir.matches(SUPPORTED_INTERNAL_DIR_NAME_MASK)) {
             throw new IOException(String.format(
@@ -225,47 +209,9 @@ public class WorkDirManager {
             directories.put(DirType.INTERNAL_DIR, internalDirFile);
 
             // Create the directory on-demand
-            internalDirPath = PathUtils.fileToPath(internalDirFile);
+            final Path internalDirPath = PathUtils.fileToPath(internalDirFile);
             Files.createDirectories(internalDirPath);
             LOGGER.log(Level.INFO, "Using {0} as a remoting work directory", internalDirPath);
-
-            var pids = internalDirPath.resolve("pids");
-            Files.createDirectories(pids);
-            var others = Files.list(pids)
-                    .map(Path::getFileName)
-                    .map(Path::toString)
-                    .filter(n -> !n.equals(agentName))
-                    .toList();
-            if (!others.isEmpty()) {
-                LOGGER.warning(() -> pids + " suggests this workDir is being used by agents other than " + agentName
-                        + " which could lead to file corruption and other issues: " + others);
-            }
-            var lockFile = pids.resolve(agentName);
-            var fc = FileChannel.open(lockFile, StandardOpenOption.CREATE, StandardOpenOption.WRITE);
-            var shouldDelete = new AtomicBoolean();
-            Runtime.getRuntime()
-                    .addShutdownHook(new Thread(
-                            () -> {
-                                try {
-                                    fc.close();
-                                } catch (IOException x) {
-                                    LOGGER.log(Level.WARNING, "failed to close " + lockFile, x);
-                                }
-                                if (shouldDelete.get()) {
-                                    try {
-                                        Files.delete(lockFile);
-                                    } catch (IOException x) {
-                                        LOGGER.log(Level.WARNING, "failed to delete " + lockFile, x);
-                                    }
-                                }
-                            },
-                            "clean up " + lockFile));
-            if (fc.tryLock() == null) {
-                throw new IOException("Another process is already running the agent: " + lockFile);
-            }
-            shouldDelete.set(true);
-            Channels.newOutputStream(fc)
-                    .write(Long.toString(ProcessHandle.current().pid()).getBytes(StandardCharsets.US_ASCII));
 
             // Create components of the internal directory
             createInternalDirIfRequired(internalDirFile, DirType.JAR_CACHE_DIR);
