+++ src/test/java/org/jenkinsci/remoting/engine/WorkDirManagerTest.java
@@ -76,7 +76,7 @@ public class WorkDirManagerTest {
 
         // Initialize and check the results
         final Path createdDir = WorkDirManager.getInstance()
-                .initializeWorkDir(dir, WorkDirManager.DirType.INTERNAL_DIR.getDefaultLocation(), false, "x");
+                .initializeWorkDir(dir, WorkDirManager.DirType.INTERNAL_DIR.getDefaultLocation(), false);
         assertThat(
                 "The initialized " + WorkDirManager.DirType.INTERNAL_DIR + " differs from the expected one",
                 createdDir,
@@ -107,7 +107,7 @@ public class WorkDirManagerTest {
 
         // Initialize and check the results
         final Path createdDir = WorkDirManager.getInstance()
-                .initializeWorkDir(workDir, WorkDirManager.DirType.INTERNAL_DIR.getDefaultLocation(), false, "x");
+                .initializeWorkDir(workDir, WorkDirManager.DirType.INTERNAL_DIR.getDefaultLocation(), false);
         assertThat(
                 "The initialized " + WorkDirManager.DirType.INTERNAL_DIR + " differs from the expected one",
                 createdDir.toFile(),
@@ -129,8 +129,7 @@ public class WorkDirManagerTest {
         final File remotingDir = new File(workDir, internalDirectoryName);
 
         // Initialize and check the results
-        final Path createdDir =
-                WorkDirManager.getInstance().initializeWorkDir(workDir, internalDirectoryName, false, "x");
+        final Path createdDir = WorkDirManager.getInstance().initializeWorkDir(workDir, internalDirectoryName, false);
         assertThat(
                 "The initialized " + WorkDirManager.DirType.INTERNAL_DIR + " differs from the expected one",
                 createdDir.toFile(),
@@ -145,7 +144,7 @@ public class WorkDirManagerTest {
         File foo = tmpDir.newFile("foo");
         try {
             WorkDirManager.getInstance()
-                    .initializeWorkDir(foo, WorkDirManager.DirType.INTERNAL_DIR.getDefaultLocation(), false, "x");
+                    .initializeWorkDir(foo, WorkDirManager.DirType.INTERNAL_DIR.getDefaultLocation(), false);
         } catch (IOException ex) {
             assertThat(
                     "Wrong exception message",
@@ -238,7 +237,7 @@ public class WorkDirManagerTest {
     public void shouldCreateLogFilesOnTheDisk() throws Exception {
         final File workDir = tmpDir.newFolder("workDir");
         final WorkDirManager mngr = WorkDirManager.getInstance();
-        mngr.initializeWorkDir(workDir, "remoting", false, "x");
+        mngr.initializeWorkDir(workDir, "remoting", false);
         mngr.setupLogging(mngr.getLocation(WorkDirManager.DirType.INTERNAL_DIR).toPath(), null);
 
         // Write something to logs
@@ -296,7 +295,7 @@ public class WorkDirManagerTest {
         if (passToManager) {
             mngr.setLoggingConfig(loggingConfigFile);
         }
-        mngr.initializeWorkDir(workDir, "remoting", false, "x");
+        mngr.initializeWorkDir(workDir, "remoting", false);
         mngr.setupLogging(mngr.getLocation(WorkDirManager.DirType.INTERNAL_DIR).toPath(), null);
 
         // Write something to logs
@@ -328,7 +327,7 @@ public class WorkDirManagerTest {
         // Initialize and check the results
         try {
             WorkDirManager.getInstance()
-                    .initializeWorkDir(workDir, WorkDirManager.DirType.INTERNAL_DIR.getDefaultLocation(), true, "x");
+                    .initializeWorkDir(workDir, WorkDirManager.DirType.INTERNAL_DIR.getDefaultLocation(), true);
         } catch (IOException ex) {
             assertThat(
                     "Unexpected exception message",
@@ -341,7 +340,7 @@ public class WorkDirManagerTest {
 
     private void assertAllocationFails(File workDir, String internalDirName) throws AssertionError {
         try {
-            WorkDirManager.getInstance().initializeWorkDir(workDir, internalDirName, false, "x");
+            WorkDirManager.getInstance().initializeWorkDir(workDir, internalDirName, false);
         } catch (IOException ex) {
             assertThat(
                     ex.getMessage(),
@@ -364,7 +363,7 @@ public class WorkDirManagerTest {
             throws AssertionError, IOException {
         WorkDirManager instance = WorkDirManager.getInstance();
         instance.disable(type);
-        instance.initializeWorkDir(workDir, "remoting", false, "x");
+        instance.initializeWorkDir(workDir, "remoting", false);
 
         // Checks
         assertThat(
@@ -403,7 +402,7 @@ public class WorkDirManagerTest {
 
         try {
             WorkDirManager.getInstance()
-                    .initializeWorkDir(dir, WorkDirManager.DirType.INTERNAL_DIR.getDefaultLocation(), false, "x");
+                    .initializeWorkDir(dir, WorkDirManager.DirType.INTERNAL_DIR.getDefaultLocation(), false);
         } catch (IOException ex) {
             assertThat(
                     "Wrong exception message for " + flag,
